Organise UmbracoViewPageTests into partials.
